refactor(codegen): extract std::io::stdin::* into stdlib/io_stdin.rs (Round 1.11)
2 fns lifted into IoStdinStdlib<'ctx>: read_line, read_line_status.

codegen.rs: 43509 -> 43433 (-76 lines).
